+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to MiniJinja are documented here.
 
 - Added `default` filter.
 - Added `startingwith` and `endingwith` tests.
+- Added global functions support.
+- Added `range` function.
+- Fixed panic caused by `super()` calls outside of blocks.
 
 # 0.5.0